Move MainDispatcherRule to utils package
Relocated MainDispatcherRule from the root test package to the utils test package and updated all imports accordingly.
