+++ app/src/test/java/org/ole/planet/myplanet/ui/voices/NewsViewModelTest.kt
@@ -12,7 +12,7 @@ import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
-import org.ole.planet.myplanet.MainDispatcherRule
+import org.ole.planet.myplanet.utils.MainDispatcherRule
 import org.ole.planet.myplanet.repository.VoicesRepository
 import org.ole.planet.myplanet.utils.DispatcherProvider
